backend/pyproject.toml
[project]
name = "lightweight-pve-panel-backend"
version = "0.1.0"
description = "Python backend for the lightweight PVE management system"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlmodel>=0.0.22,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "cryptography>=44.0.0,<45.0.0",
  "websockets>=15.0.1,<16.0.0",
  "pytest>=8.3.3,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
